requirements.txt
asgiref==3.8.1
attrs==23.2.0
chardet==5.2.0
Django==5.0.4
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docx2pdf==0.1.8
drf-spectacular==0.27.2
et-xmlfile==1.1.0
fpdf==1.7.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
openpyxl==3.1.2
pillow==10.3.0
PyJWT==2.8.0
PyPDF2==3.0.1
PyYAML==6.0.1
referencing==0.34.0
reportlab==4.1.0
rpds-py==0.18.0
sqlparse==0.5.0
tqdm==4.66.2
uritemplate==4.1.1
